Got rid of some duplicated code
